Rename package according to guidelines - Update code namespace - Update as-build artifact naming - Update documentation references
