+++ pyproject.toml
@@ -33,7 +33,7 @@ deps =
 commands = pytest ./tests
 
 [testenv:coverage]
-commands = pytest --cov=ansys.grantami.common ./tests
+commands = pytest --cov=ansys.openapi.common ./tests
 
 [testenv:lint]
 deps =
