+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = openapi_client_common
-version = attr: ansys.grantami.common.__version__
-description = Provides a helper to create sessions for use with Ansys Granta MI API clients.
+name = ansys_openapi_common
+version = attr: ansys.openapi.common.__version__
+description = Provides a helper to create sessions for use with Ansys OpenAPI clients.
 long_description = file: README.md, LICENSE
-keywords = Ansys, Granta MI
+keywords = Ansys, OpenAPI
 license = MIT
 classifiers =
     License :: OSI Approved :: MIT License
