+++ src/ansys/openapi/common/__init__.py
@@ -1,5 +1,3 @@
-import logging
-
 from ._session import ApiClientFactory
 from ._util import SessionConfiguration
 from ._exceptions import ApiConnectionException, ApiException, AuthenticationWarning
