+++ src/ansys/openapi/common/_exceptions.py
