+++ src/ansys/openapi/common/_oidc.py
