+++ src/ansys/openapi/common/_session.py
