+++ src/ansys/openapi/common/_util.py
